refactor nwm loading to pull out common code, add ability to choose nwm ver for retro.
